add main to a dedicated file
